requirements.txt
Logbook==1.0.0
Flask==0.11.1
lxml==3.6.4
redis==2.10.5
requests==2.11.1
six==1.10.0
Pillow==3.3.1
celery==3.1.23
Flask-Script==2.0.5
git+https://github.com/slushkovsky/usefutils.git